fix: 2주차 피드백 수정 코드로 변경
